made map a separate page
