+++ app/assets/stylesheets/map.css
@@ -1,0 +1,13 @@
+#map {
+    height: 100%;
+  }
+  
+  .navbar-fixed{
+    position: fixed;
+  }
+  
+  html, body {
+    height: 100%;
+    margin: 0;
+    padding: 0;
+  }
