+++ app/views/static_pages/dashboard.html.erb
@@ -7,13 +7,15 @@
   </head>
 
   <body>
+  <% provide(:nav_text, "Giving Gardens - Administrative Dashboard") %>
+  <!--
     <div class="navbar-fixed">
       <nav>
         <div class="nav-wrapper light-green">
           <a href="./" class="brand-logo">Giving Gardens - Administrative Dashboard</a>
         </div>
       </nav>
-    </div>
+    </div> -->
 
     <div class="container">
       <h3>Gardens</h3>
