+++ app/views/static_pages/home.html.erb
@@ -1,17 +1,4 @@
 <%= stylesheet_link_tag "home" %>
 <%= javascript_pack_tag 'home-pack' %>
-<!--navbar-->
-<div class="navbar-fixed">
-  <nav>
-    <div class="nav-wrapper light-green">
-      <a href="#" class="brand-logo" style="float: left; margin-left: .5vw;">Giving Gardens</a>
-      <a href="/produce-by-garden" style="float: right;">Produce by Garden</a>
-      <a href="/dashboard" style="float: right;">Dashboard Temp Link</a>
-    </div>
-  </nav>
-</div>
-<!--navbar-->
+<% provide(:nav_text, "Giving Gardens") %>
 
-<div id="map"></div>
-<script id="GoogleMapsScript" src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBcxjlc7QhJOAskOrvlDZubGVNdT4qDxck&callback=initMap"></script>
-<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
