+++ app/views/static_pages/map.html.erb
@@ -1,0 +1,6 @@
+<%= stylesheet_link_tag "map" %>
+<%= javascript_pack_tag 'map-pack' %>
+<% provide(:nav_text, "Giving Gardens Map") %>
+<div id="map"></div>
+<script id="GoogleMapsScript" src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBcxjlc7QhJOAskOrvlDZubGVNdT4qDxck&callback=initMap"></script>
+<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
